requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Deep Learning
torch==2.1.0
torchvision==0.16.0
efficientnet-pytorch==0.7.1

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Database
pymongo==4.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# PDF Generation
reportlab==4.0.7

# CORS
python-cors==1.0.0